Moved map in a component
